requirements.txt
Django>=3.2,<4.0
django-simple-history>=2.12.0
celery>=5.2.0
django-celery-beat>=2.2.0
